+++ package.json
@@ -2,14 +2,14 @@
   "name": "yddict",
   "version": "1.1.0",
   "description": "A CLI tool for querying youdao dict.",
-  "main": "./index.js",
+  "main": "./bin/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "kenshinji <gundam0083ster@gmail.com>",
   "license": "MIT",
   "bin": {
-    "yd": "./index.js"
+    "yd": "./bin/index.js"
   },
   "dependencies": {
     "chalk": "^2.1.0",
